frontend/requirements.txt
# requirements.txt (frontend)
gradio==4.19.2
requests==2.32.3
pillow==11.2.1
pandas==2.1.4
python-dotenv==1.0.0